Fixed tests in remote_control
We're not kicking off a build when remote control starts any more so
the tests were waiting on a message that wouldn't be sent.
Additionally, since we're calling `Code.eval_string` rather than
`Code.compile_string`, a successful compilation won't result in a
`module_updated` message any more. This was an actual bug.