Gated in_mix_project on builds
When we're doing a project-wide build, that will inevitable recompile
the mix.exs, but in_mix_project requires a certain version of the
mix.exs to be defined in memory. If it's not the most recent one, we
get an exception.

Gating them via a build lock means we don't get this conflict any more