Build throttling for compile-as-you-type (#61)
* Build throttling for compile-as-you-type

Added throttling for compile-as-you-type. Moved most of the
functionality out of `build.ex` and into `state.ex`. Then implemented
tick-based throttling. The result is that if a file has been quiet for
250ms, we compile it. This makes the builds feel buttery smooth.


The remote control app was not picking up its config (it can't, it's
running in the context of the app unde test), so I passed the config
in. This allows us to adjust the throttle timeouts for tests.

This change saved 12 seconds under test.

* Removed build directory caching

This was causing a ton of warnings of modules being redefined. I think
caching the build directory could have worked if we could exclude our
applications, which would effectively cache our deps, but I could not
find any documentation indicating that this was possible.