Converted positions to be one-based (#87)
* Converted positions to be one-based

When I designed lexical, I thought it would be good to have zero-based
positions like the LS protocol and elixir's conventions. I thought
that getting the compiler to emit zero-based messages would be as
simple as adding a compiler setting. However, it's not so easy. In
fact, it's impossible at the present moment.

So, we can see the code is starting to get a lot of `line + 1` and
`line - 1` kind of code, which stinks.

Combined with the last commit, this changes source files, code units,
positions and ranges to be one-based, and we convert from the
zero-based language server at the boundaries. Now, the only code that
adds or subtracts will be string manipulation code, which we'll likely
put behind some kind of interface.

Get ready for a series of off-by-one errors