Improved struct completion (#181)
Struct completion under elixir sense was extremely inconsistent from
context to context and even file to file. This was extremely
frustrating, as structs might or might not complete depending,
seemingly on the time of day.

On the other hand, module completion was pretty consistent, so to fix
the issue, prior to sending a document for completion, if we're in a
struct reference, we strip it out to get the original module reference
back. This means we'll only get module completions from elixir_sense,
but we already check for struct references and if we detect that we're
in one, turn module completions into struct completions.

I've played around with this and it does seem to fix the problems I
was having prior.