We now use Lexical.Plugin.Diagnostics (#197)
* We now use Lexical.Plugin.Diagnostics

Switched from using Mix.Compiler.Diagnostic to our plugin
diagnostics. This is so we can handle diagnostics from everywhere the
same way.

While I was converting error.ex, I found that we went through great
lengths to extract the file path from the error messages, but we
actually had it right next to us in the build, so I passed in the
document, which eliminated some complexity there.


* Handled the case where exceptions were blowing up errors

We had some bandaid code in error.ex and diagnostics.ex that
specifically handled Mix.Error escaping being turned into
diagnostics. This was because the "message info" can be an
exception. Added a function head to fix this problem.

* Added missed case to handle mix errors

Fixes #198