Indexing features for modules (#347)
* Indexing features for modules

This commit introduces indexing for module references. It currently
will only work for files to which we have the source code, turns them
into AST and then applies one or more "extractors" on it. An extractor
runs over AST and indetifies things it's looking for, and then emits
entries, which will be used to build a searchable index.

Currently, there is one extractor implemented: Module references. This
does a fairly exhaustive search for module references, wherever
they're hiding. It then uses the aliases module to find the original
source module.

Note that most of the higher level functions are just placeholders. An
indexing strategy needs to find things inside of .exs files as well as
.ex files, and we currently have no way to index core elixir (this
will be highly dependent on how elixir was installed and if we can get
the source code to the version we're running on).

* Performance fixes / code improvements

Turns out `Aliases.at()` was super slow --especially since we called
it so much when resolving aliases during indexing. This was due to it
having to re-parse the document multiple times for each alias
reference. Moving to sourceror and allowing the AST to be sent in cut
the indexing time for lexical and all its deps from 60 seconds to 10.


* ETS based storage / Fuzzy matching

This commit has ETS based file storage and a fuzzy matching system.
I was having a lot of trouble coming up with an efficient way to perform
fuzzy matching against a corpous of subjects efficiently using the data
stores. ETS is... ets, and it appears that it's impossible to use it for
any type of detailed string matching. SQLite had similar problems. While
it had a lot more complicated and helpful utilities, it just couldn't do
what I wanted when it comes to fuzziness.

Then I thought, hey, why not just store the subjects in memory with all
the places that they appear? Then you just keep that up to date, and you
remove a lot of duplication, which improves performance a lot.

Prior to this, ETS was taking around 300ms to do simple-ish matching,
and SQLite was taking around 180ms. Fuzzy can do it in 14ms. Better yet,
building a fuzzy from scratch only takes 40ms for lexical and all its
deps. This includes doing much more interesting matching.

I think we have a winner.