General and VS Code-specific completions improvements (#410)
This commit improves completions for the general case and adds special handling when the client is VS Code that fixes numerous issues related to completions in that editor.

General improvements:

* Partially revert #398, which introduced a UX regression that caused slow/high latency completions.
* Don't insert additional, unnecessary newlines with multi-line snippets.
* Improved snippets for `defdelegate`, `defguard`, `defguardp`, `with`.
* Consistent capitalization and naming in completion labels.
* Fix a bug that caused a do/end block completion to be proposed when the cursor was after `@do`, which would occur whenever typing `@doc`.

VS Code:

VS Code has some problematic behavior when dealing with empty completion lists. When the client is VS Code, we now attempt to never send empty completion lists if it is possible that further typing will result in some completions. Instead, we send a larger list up front and allow the editor to filter.

See #400 for more context.

---------

Co-authored-by: Scott Ming <therealscottming@gmail.com>