Find function references (#516)
* Find references: Functions

Implemented find references for functions.

Initially, very few references were found, and this was because we
were deduping the index by id. Removing this call allows us to see the
entire project's references, with the notable exception of imported
macros. I believe we'll need to augment the extractor with a
compilation tracer if we wish to see those. The bad news is lexical
takes around 12 seconds to compile, but around 3 seconds to index.


* Tightened up indexing

The extractors were slightly wonky when extracting ranges. They would
leave off the trailing paren in calls, and would capture the whole
function definition. This fixes both issues.

There are a couple of range issues that remain, but this is a marked improvement.

* Indexer: Handled function references in pipelines