Block hierarchy (#541)
* Added the notion of structure to the index

We're emitting entries sparsely now, so it's quite possible that the
structure of an entry won't be represented. To fix this, I made a
special entry type that represents the block structure of
documents. This can then be queried to get parents or siblings of a
given entry.

This will be useful when we implement module attributes so we can
figure out their scoping.