Workspace symbols (#674)
* Workspace symbols

Wired up workspace symbols. Things work, but this needs some
_love_. The search is kind of slow, and results are not particularly
relevant. I also don't use this feature much, so I don't know if we're
supposed to find _everything_ or, say, definitions.

I'm putting this up here as a draft PR while I open another branch
that will work toward improving relevance and performance.

* Improved performance / utility of fuzzy matcher and scorer

Made somewhat significant changes to the fuzzy matcher to improve
performance and relevance of the results.

I found that the matcher was returning a ton of results when it was
returning references to things, so I pared it down to only consider
definitions, which seems to improve the relevance of results a lot.

I also only had it return definitions in one of the current project's
applications, as searching your dependencies for a definition didn't
seem to be a common use case.

The fuzzy matcher also now considers the type of the entry being
indexed when it builds the subject. This gives us much nicer and more
relevant search results, and gives us the ability to tune them more
specifically in the future.