Improved store error handling (#737)
The store has a variety of states, and can return errors when it is
not enabled. In light of this, we should return ok tuples when it is
successful to keep the api consistent and make callers aware that they
need to handle errors.

In addition, if loading / enabling takes a while, the store can time
out when it's being queried, so unless it's enabled, queries will
return empty results, and the enabled check uses persistent term.