Indexed delegated functions (#729)
* Indexed delegated functions

Functions created with defdelegate weren't being indexed, so they
weren't showing up in find references, document symbols,  or go to
definition.

This change emits two entries when it encounters a `defdelegate` call,
one for the local function definition in the calling module, and a
function reference to the delegated module. One possibly strange thing
is that the function that's being created by defdelegate shows up as
`def` in document symbols. This is technically correct, you're
defining a public function, but it might be confusing in the UI.

Fixes #724