[feat] Group functions by name + type + arity in document symbols (#833)
If a module had multiple heads for a function that only differed in
parameter pattern matches, document symbols was very cluttered. This
change groups functions in a module by their name and arity, with each
implementation nested in that grouping. This allows editors to
collapse function groups to save space.

This change also removes any whitespace or linebreaks from function names.